refactor: move path functions to path module
