+++ server_manager/src/properties_manager.py
@@ -2,39 +2,18 @@
 especially `online-mode`.
 """
 
-import os
-import re
-import sys
 from functools import lru_cache
 from pathlib import Path
+import re
+import sys
 
 from colorama import Fore
 
 from .exceptions import InvalidServerStateError
-
+from .paths import get_server_path
 from .utils import bool2str, str2bool
 
 ONLINE_MODE_PATTERN: re.Pattern = re.compile(r"(online-mode=)(\w+)", re.IGNORECASE)
-DATA_PATH = Path(__file__).parent.joinpath("data/server-path.txt")
-os.makedirs(DATA_PATH.parent, exist_ok=True)
-
-
-def get_server_path() -> Path:
-    """Returns the server path. If it not stored, it will ask the user and then
-    store the response after validating it.
-
-    Returns:
-        Path: root dir of the server (folder which contains `server.properties`).
-    """
-
-    if DATA_PATH.is_file():
-        return Path(DATA_PATH.read_text().strip())
-
-    server_path = input("Write the path of the server: ").strip()
-    validate_server_path(server_path)
-
-    DATA_PATH.write_text(Path(server_path).as_posix())
-    return Path(server_path)
 
 
 def validate_server_path(server_path: str):
